5단계 - 자동차 경주(리팩토링) | refactor : 테스트 코드들도 MVC 패턴으로 패키지 분리
