+++ src/test/java/racinggame/domain/TraceTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.Test;
 import racinggame.domain.Trace;
