+++ src/test/java/racinggame/view/InputValidatorTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.view;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
